fast5_fastq_converter/requirements.txt
google-cloud-storage==1.13.0
google-cloud-pubsub==0.38.0